working mobile menu search functionality
