+++ src/App.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import Footer from './components/Footer';
-import Header from './components/Header';
 import GlobalStyle from './GlobalStyle';
 import Routes from './Routes';
 
@@ -10,12 +9,9 @@ function App() {
 		<>
 			<GlobalStyle />
 			<Router>
-				<Header />
 				<Routes />
 				<Footer />
-
 			</Router>
-
 		</>
 	);
 }
